app/requirements.txt
kafka-python
psycopg2

pandas


